requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
apscheduler==3.10.4
aiohttp==3.9.5
python-dotenv==1.0.1
